Revert "Health and Verify Setup"
