--- .aws-sam/build.toml
@@ -1,20 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions]
-[function_build_definitions.c6feb298-2bea-4e7c-b809-b5b6ed70806d]
-codeuri = "G:\\identity-service\\health"
-runtime = "go1.x"
-architecture = "x86_64"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["HealthFunction"]
-
-[function_build_definitions.dceba13d-8078-4d69-b956-99d0ef88a1dc]
-codeuri = "G:\\identity-service\\verify"
-runtime = "go1.x"
-architecture = "x86_64"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["VerifyFunction"]
-
-[layer_build_definitions]
