--- health/go.mod
@@ -1,5 +1,0 @@
-require github.com/aws/aws-lambda-go v1.23.0
-
-module health
-
-go 1.16
