--- samconfig.toml
@@ -1,12 +1,0 @@
-version = 0.1
-[default]
-[default.deploy]
-[default.deploy.parameters]
-stack_name = "identity-service"
-s3_bucket = "aws-sam-cli-managed-default-samclisourcebucket-7idfnjoy8ob3"
-s3_prefix = "identity-service"
-region = "us-east-2"
-confirm_changeset = true
-capabilities = "CAPABILITY_IAM"
-disable_rollback = true
-image_repositories = []
