--- verify/go.mod
@@ -1,8 +1,0 @@
-module main
-
-go 1.17
-
-require (
-	github.com/aws/aws-lambda-go v1.27.1
-	golang.org/x/crypto v0.0.0-20211215153901-e495a2d5b3d3
-)
